splits FeedCard
