+++ src/pages/Feed/index.jsx
@@ -1,24 +1,17 @@
-import React from 'react'
-import FeedCard from '../../components/FeedCard'
-import MainHeader from '../../components/MainHeader'
-import './style.css'
+import React from "react";
+import FeedCard from "../../components/FeedCard";
+import MainHeader from "../../components/MainHeader";
+import "./style.css";
 
 export default function Feed() {
-    return (
-        <div>
-            <MainHeader />
-            <div className='feedCard'>
-            <FeedCard className=''/>
-            <FeedCard />
-            <FeedCard />
-            <FeedCard />
-            <FeedCard />
-            <FeedCard />
-            <FeedCard />
-            <FeedCard />
-            <FeedCard />
-            <FeedCard />
-            </div>
-        </div>
-    )
+  return (
+    <div>
+      <MainHeader />
+      <div className="feedCard">
+        {[...Array(10)].map(() => (
+          <FeedCard />
+        ))}
+      </div>
+    </div>
+  );
 }
